Move ExpertModeModal component to pures
