+++ src/cow-react/common/pure/ExpertModeModal/index.cosmos.tsx
@@ -1,0 +1,13 @@
+import { ExpertModeModal, ExpertModeModalProps } from './index'
+
+const defaultProps: ExpertModeModalProps = {
+  isOpen: true,
+  onDismiss() {
+    console.log('Dismiss')
+  },
+  onEnable() {
+    console.log('Enable')
+  },
+}
+
+export default <ExpertModeModal {...defaultProps} />
