+++ src/custom/components/Settings/SettingsMod.tsx
@@ -1,10 +1,10 @@
 // eslint-disable-next-line no-restricted-imports
-import { t, Trans } from '@lingui/macro'
+import { Trans } from '@lingui/macro'
 // import { Percent } from '@uniswap/sdk-core'
 // import { useWeb3React } from '@web3-react/core'
 // import { AUTO_ROUTER_SUPPORTED_CHAINS } from 'lib/hooks/routing/clientSideSmartOrderRouter'
 import { useCallback, useContext, useRef, useState } from 'react'
-import { Settings, X } from 'react-feather'
+import { Settings } from 'react-feather'
 // import ReactGA from 'react-ga4'
 import { Text } from 'rebass'
 import styled, { ThemeContext } from 'styled-components/macro'
@@ -13,9 +13,7 @@ import { useModalIsOpen, useToggleSettingsMenu } from 'state/application/hooks'
 import { ApplicationModal } from 'state/application/reducer'
 import { useExpertModeManager, useRecipientToggleManager } from 'state/user/hooks'
 import { ThemedText } from 'theme'
-import { ButtonError } from 'components/Button'
 import { AutoColumn } from 'components/Column'
-import Modal from 'components/Modal'
 import QuestionHelper from 'components/QuestionHelper'
 import { RowBetween, RowFixed } from 'components/Row'
 import Toggle from 'components/Toggle'
@@ -29,6 +27,7 @@ import {
 
 // MOD imports
 import { SettingsTabProp } from '.'
+import { ExpertModeModal } from '@cow/common/pure/ExpertModeModal'
 
 export const StyledMenuIcon = styled(Settings)`
   height: 20px;
@@ -40,18 +39,6 @@ export const StyledMenuIcon = styled(Settings)`
 
   :hover {
     opacity: 0.7;
-  }
-`
-
-const StyledCloseIcon = styled(X)`
-  height: 20px;
-  width: 20px;
-  :hover {
-    cursor: pointer;
-  }
-
-  > * {
-    stroke: ${({ theme }) => theme.text1};
   }
 `
 
@@ -109,21 +96,6 @@ export const MenuFlyout = styled.span`
   user-select: none;
 `
 
-const Break = styled.div`
-  width: 100%;
-  height: 1px;
-  background-color: ${({ theme }) => theme.bg3};
-`
-
-export const ModalContentWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 2rem 0;
-  background-color: ${({ theme }) => theme.bg2};
-  border-radius: 20px;
-`
-
 export default function SettingsTab({ className, placeholderSlippage, SettingsButton }: SettingsTabProp) {
   // const { chainId } = useWeb3React()
 
@@ -170,46 +142,14 @@ export default function SettingsTab({ className, placeholderSlippage, SettingsBu
   return (
     // https://github.com/DefinitelyTyped/DefinitelyTyped/issues/30451
     <StyledMenu ref={node as any} className={className}>
-      <Modal isOpen={showConfirmation} onDismiss={() => setShowConfirmation(false)} maxHeight={100}>
-        <ModalContentWrapper>
-          <AutoColumn gap="lg">
-            <RowBetween style={{ padding: '0 2rem' }}>
-              <div />
-              <Text fontWeight={500} fontSize={20}>
-                <Trans>Are you sure?</Trans>
-              </Text>
-              <StyledCloseIcon onClick={() => setShowConfirmation(false)} />
-            </RowBetween>
-            <Break />
-            <AutoColumn gap="lg" style={{ padding: '0 2rem' }}>
-              <Text fontWeight={500} fontSize={20}>
-                <Trans>
-                  Expert mode turns off the confirm transaction prompt and allows high slippage trades that often result
-                  in bad rates and lost funds.
-                </Trans>
-              </Text>
-              <Text fontWeight={600} fontSize={20}>
-                <Trans>ONLY USE THIS MODE IF YOU KNOW WHAT YOU ARE DOING.</Trans>
-              </Text>
-              <ButtonError
-                error={true}
-                padding={'12px'}
-                onClick={() => {
-                  const confirmWord = t`confirm`
-                  if (window.prompt(t`Please type the word "${confirmWord}" to enable expert mode.`) === confirmWord) {
-                    toggleExpertMode()
-                    setShowConfirmation(false)
-                  }
-                }}
-              >
-                <Text fontSize={20} fontWeight={500} id="confirm-expert-mode">
-                  <Trans>Turn On Expert Mode</Trans>
-                </Text>
-              </ButtonError>
-            </AutoColumn>
-          </AutoColumn>
-        </ModalContentWrapper>
-      </Modal>
+      <ExpertModeModal
+        isOpen={showConfirmation}
+        onDismiss={() => setShowConfirmation(false)}
+        onEnable={() => {
+          toggleExpertMode()
+          setShowConfirmation(false)
+        }}
+      />
       <SettingsButton expertMode={expertMode} toggleSettings={toggle} />
       {/* <StyledMenuButton onClick={toggle} id="open-settings-dialog-button">
         <StyledMenuIcon />
